backend/requirements.txt
Django
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
whitenoise==6.6.0
